Fix/aimsdelete

+++ plugins/EpadDB.js
@@ -7953,13 +7953,49 @@ async function epaddb(fastify, options, done) {
       )
         reply.send(new UnauthorizedError('User is not admin, cannot delete from system'));
       else {
-        const aimDelete = await fastify.deleteAimsInternal(
-          request.params,
-          request.epadAuth,
-          request.query,
-          request.body
-        );
-        reply.code(200).send(aimDelete);
+        // for each aim check if the user has a right to delete
+        let aimsThatCanBeDeleted = [];
+        const aimsThatCannotBeDeleted = [];
+        if (request.body && Array.isArray(request.body)) {
+          for (let i = 0; i < request.body.length; i += 1) {
+            if (
+              request.epadAuth.admin ||
+              fastify.isOwnerOfProject(request, request.params.project) ||
+              // eslint-disable-next-line no-await-in-loop
+              (await fastify.isCreatorOfObject(request, {
+                level: 'aim',
+                objectId: request.body[i],
+                project: request.params.project,
+              })) === true
+            )
+              aimsThatCanBeDeleted.push(request.body[i]);
+            else aimsThatCannotBeDeleted.push(request.body[i]);
+          }
+        } else aimsThatCanBeDeleted = request.body;
+        if (!Array.isArray(aimsThatCanBeDeleted) || aimsThatCanBeDeleted.length > 0) {
+          const aimDelete = await fastify.deleteAimsInternal(
+            request.params,
+            request.epadAuth,
+            request.query,
+            aimsThatCanBeDeleted
+          );
+          if (aimsThatCannotBeDeleted.length > 0)
+            new EpadNotification(
+              request,
+              `Only some of the aims were deleted. User does not have the sufficient rights to delete `,
+              aimsThatCannotBeDeleted.join(', ')
+            ).notify(fastify);
+          reply.code(200).send({ message: aimDelete, aimsThatCannotBeDeleted });
+        } else {
+          reply.send(
+            new InternalError(
+              `Aims ${JSON.stringify(request.body)}  deletion from project ${
+                request.params.project
+              }`,
+              new Error('User does not have sufficient rights')
+            )
+          );
+        }
       }
     } catch (err) {
       reply.send(
